requirements.txt
easyocr
cv
transformers
datasets
numpy
pandas
torch
pyenchant